backend/requirements.txt
PyJWT~=2.8.0
Django~=5.0.3
djangorestframework~=3.15.0
redis~=5.0.3
djangorestframework-simplejwt
django-redis
drf-yasg